Fix: import/no-cycle
